+++ plugin-document/src/test/java/nl/xillio/xill/plugins/mongodb/services/BsonValueConverterTest.java
@@ -19,17 +19,19 @@ import nl.xillio.xill.api.components.MetaExpression;
 import nl.xillio.xill.api.data.Date;
 import nl.xillio.xill.api.data.DateFactory;
 import nl.xillio.xill.plugins.mongodb.data.MongoObjectId;
+import nl.xillio.xill.plugins.mongodb.data.MongoRegex;
 import nl.xillio.xill.plugins.mongodb.services.serializers.BinarySerializer;
+import nl.xillio.xill.plugins.mongodb.services.serializers.MongoRegexSerializer;
 import nl.xillio.xill.plugins.mongodb.services.serializers.ObjectIdSerializer;
 import nl.xillio.xill.services.json.JacksonParser;
 import org.apache.commons.io.IOUtils;
 import org.bson.*;
-import org.bson.types.Binary;
 import org.bson.types.ObjectId;
 import org.testng.annotations.Test;
 
 import java.time.Instant;
 import java.util.Collections;
+import java.util.regex.Pattern;
 
 import static nl.xillio.xill.api.components.ExpressionBuilder.*;
 import static org.mockito.Matchers.eq;
@@ -39,10 +41,9 @@ import static org.testng.Assert.assertNotNull;
 
 
 public class BsonValueConverterTest {
-
     @Test
     public void testConvert() throws Exception {
-        BsonValueConverter converter = new BsonValueConverter(null, null, null);
+        BsonValueConverter converter = new BsonValueConverter(null, null, null, null);
         JacksonParser parser = new JacksonParser(false);
 
         assertEquals(converter.convert(new BsonString("Hello")), fromValue("Hello"));
@@ -57,12 +58,12 @@ public class BsonValueConverterTest {
 
 
     @Test
-    public void testConvertDates() throws Exception {
+    public void testConvertDates() {
         DateFactory dateFactory = mock(DateFactory.class);
         Date result = mock(Date.class);
         when(dateFactory.from(any())).thenReturn(result);
 
-        BsonValueConverter converter = new BsonValueConverter(dateFactory, null, null);
+        BsonValueConverter converter = new BsonValueConverter(dateFactory, null, null, null);
 
         MetaExpression timestampDate = converter.convert(new BsonTimestamp(1000, 0));
         assertEquals(timestampDate.getMeta(Date.class), result);
@@ -74,9 +75,9 @@ public class BsonValueConverterTest {
     }
 
     @Test
-    public void testConvertObjectId() throws Exception {
+    public void testConvertObjectId() {
         ObjectIdSerializer objectIdSerializer = new ObjectIdSerializer();
-        BsonValueConverter converter = new BsonValueConverter(null, objectIdSerializer, null);
+        BsonValueConverter converter = new BsonValueConverter(null, objectIdSerializer, null, null);
         ObjectId objectId = new ObjectId();
         MetaExpression expression = converter.convert(new BsonObjectId(objectId));
         assertNotNull(expression.getMeta(MongoObjectId.class));
@@ -84,9 +85,19 @@ public class BsonValueConverterTest {
     }
 
     @Test
+    public void testConvertMongoRegex() {
+        MongoRegexSerializer mongoRegexSerializer = new MongoRegexSerializer();
+        BsonValueConverter converter = new BsonValueConverter(null, null, mongoRegexSerializer, null);
+        MongoRegex mongoRegex = new MongoRegex(Pattern.compile(".*"));
+        MetaExpression expression = converter.convert(new BsonRegularExpression(".*"));
+        assertNotNull(expression.getMeta(MongoRegex.class));
+        assertEquals(expression.getMeta(MongoRegex.class).getPattern().toString(), mongoRegex.getPattern().toString());
+    }
+
+    @Test
     public void testConvertBinary() throws Exception {
         BinarySerializer binarySerializer = new BinarySerializer();
-        BsonValueConverter converter = new BsonValueConverter(null, null, binarySerializer);
+        BsonValueConverter converter = new BsonValueConverter(null, null, null, binarySerializer);
         BsonBinary binary = new BsonBinary("test".getBytes());
         MetaExpression expression = converter.convert(binary);
         assertEquals(IOUtils.toString(expression.getBinaryValue().getInputStream()), "test");